requirements.txt
boto3>=1.26.0
Pillow>=10.0.0
pytest>=8.0.0
moto[s3,sts,sns]>=5.3.0
flake8>=6.0.0